% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut-internal.R
\name{donut-internal}
\alias{donut-internal}
\alias{method1_function}
\alias{method2_function}
\title{Internal donut functions}
\usage{
method1_function(data, query, k, torus, ranges, fn, ...)

method2_function(data, query, k, torus, ranges, fn, ...)
}
\description{
Internal donut functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
