% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_wrap.R
\name{plot.nn2torus}
\alias{plot.nn2torus}
\title{Plot diagnostics for an nn2torus object}
\usage{
\method{plot}{nn2torus}(x, ...)
}
\arguments{
\item{x}{an object of class \code{c("nn2torus")}, a result of
a call to \code{\link{nn2torus}}.}

\item{...}{Further arguments to be passed to \code{\link[graphics]{plot}},
\code{\link[graphics]{lines}} or \code{\link[graphics]{points}}.}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for an objects of class \code{c("nn2torus")}.
}
\section{Examples}{

See the examples in \code{\link{nn2torus}}.
}

\seealso{
\code{\link{set_val_data}}: to set validation data.
}
