% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_wrap.R
\name{plot.nnt}
\alias{plot.nnt}
\title{Plot diagnostics for an nnt object}
\usage{
\method{plot}{nnt}(x, ...)
}
\arguments{
\item{x}{an object of class \code{c("nnt")}, a result of
a call to \code{\link{nnt}}.}

\item{...}{Further arguments to be passed to \code{\link[graphics]{plot}},
\code{\link[graphics]{lines}} or \code{\link[graphics]{points}}.}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for an objects of class \code{c("nnt")}.
}
\section{Examples}{

See the examples in \code{\link{nnt}}.
}

\seealso{
\code{\link{set_val_data}}: to set validation data.
}
